mov test_access_monetary_related into test_fields_numeric
