+++ odoo/addons/test_orm/tests/__init__.py
@@ -1,6 +1,5 @@
 from . import (
     test_access_feedback,
-    test_access_monetary_related,
     test_acl,
     test_action_bindings,
     test_all_models,
